Ignore DS_Store
Added a wildcard for ignoring Mac directory meta data files.

Mac users NB : The DS_Store for the project root cannot be captured in the ignore, since it is the same level the hierarchial .gitignore, which effects  subdirectories.

The final line, '.DS_Store' SHOULD ignore the DS_Store within the same level as the ignore, but it doesn't.

This single remaining DS_Store will have to be manually unchecked in your changelists, although it is already comitted.

The repo does not contain any DS_Store files other than this topmost one.