Fixed Movement Function
Since Input Events only get called when Inputs change or are canceled the Old move function would only be called for 1 frame, but would have been required to be called in  Fixed Update.

Now when the Input Event is called we set the velocity of the Rigidbody to our direction and when it is canceled we set the velocity to 0.

Also added some more Comment Notes.