Enable/Disable Movement
Added the Ability to Enable and Disable Player Movement Controlls.
This could be used for when you are talking to an NPC so you do not accidentally move away.

I added 4 different functions for the reason that it is always good to have multiple ways to do it and it makes it hopefully easier for everyone to know what does what.