Improved animations
Background was changed from white to black, to be more subtle and less blinding

The intro text colour is no longer a direct contrast to the background, it's a more subtle grey

All animations are now shorter. Forcing the user to wait 5 seconds for everything, even exiting, it was too long.

Removed barrel roll animations, they're too over the top and nauseating for mainstream release. Instead, it's been replaced with a more subtle fade / zoom -in