Trail and Error 1
The bag has something to do with these Unity if statments according to people who have fixed the bug years ago, but this bug has evolved. I have no idea what I'm doing but I'm going to try diffrent things to fix the if statment to see if anything fixes the bug.  Here we go!